Rename getMembers file to get-members
